Refactor JUnit5 test framework provider to become more modular
Instead of the base and the vintage JUnit5 provider being two separate
entities, the vintage provider is now treated as an extension of the
base provider. This extension is only contributed, if JUnit4 is in the
classpath of the test bundle.

The test engine for Jupiter/Vintage tests are contributed via SPI. Given
that both fragments have the osgibooter as a host, this mechanism still
work, because both fragments have the same classloader. This is also the
reason why it is no longer necessary to include the base JUnit5 jars in
the vintage bundle.

This change is a prerequisite in order to support JUnit5 test suites,
which use a similar mechanism, in order to avoid having to create a
provider for each combination of Vintage, Suite and Jupiter test engine.

The change has been adjusted for JUnit5.7 and onwards. This test is
covered by JUnit5Test.testJUnit4and54Runner().